# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Secure Online Election Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /elections:
    get:
      operationId: listElections
      tags: [elections]
      summary: List elections
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: creatorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Elections list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionListResponse"
    post:
      operationId: createElection
      tags: [elections]
      summary: Create a new election (creator only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Election"
  /elections/{id}:
    get:
      operationId: getElection
      tags: [elections]
      summary: Get election by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Election
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Election"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateElection
      tags: [elections]
      summary: Update election
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElectionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Election"
    delete:
      operationId: deleteElection
      tags: [elections]
      summary: Delete election
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /elections/{id}/publish:
    post:
      operationId: publishElection
      tags: [elections]
      summary: Publish election
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Election"
  /elections/{id}/suspend:
    post:
      operationId: suspendElection
      tags: [elections]
      summary: Suspend election (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Election"
  /elections/{id}/finalize-voters:
    post:
      operationId: finalizeVoters
      tags: [elections]
      summary: Finalize voter list and lock it
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Voters finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── CANDIDATES ────────────────────────────────────────────────────────────
...
  /elections/{electionId}/results:
    get:
      operationId: getElectionResults
      tags: [results]
      summary: Get live/final election results
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Election results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionResults"

  # ─── DASHBOARD ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ─── AUTH ────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, fullName, email, role, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, election_creator, voter]
        status:
          type: string
          enum: [active, blocked, pending]
        orgName:
          type: ["string", "null"]
        electionPurpose:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Election:
      type: object
      required: [id, creatorId, title, status, createdAt]
      properties:
        id:
          type: integer
        creatorId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        registrationDeadline:
          type: ["string", "null"]
          format: date-time
        maxVoters:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, published, active, completed, suspended]
        votersFinalized:
          type: boolean
        createdAt:
          type: string
          format: date-time
        creator:
          $ref: "#/components/schemas/User"
        candidateCount:
          type: ["integer", "null"]
        voterCount:
          type: ["integer", "null"]
        voteCount:
          type: ["integer", "null"]
    ElectionInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        registrationDeadline:
          type: ["string", "null"]
          format: date-time
        maxVoters:
          type: ["integer", "null"]
    ElectionUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        registrationDeadline:
          type: ["string", "null"]
          format: date-time
        maxVoters:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, published, active, completed, suspended]
    ElectionListResponse:
      type: object
      required: [elections, total, page, limit]
      properties:
        elections:
          type: array
          items:
            $ref: "#/components/schemas/Election"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── CANDIDATE ───────────────────────────────────────────────────────────
...
    CandidateResult:
      type: object
      required: [candidateId, name, voteCount, percentage]
      properties:
        candidateId:
          type: integer
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        designation:
          type: ["string", "null"]
        voteCount:
          type: integer
        percentage:
          type: number
    ElectionResults:
      type: object
      required: [electionId, totalVotes, totalVoters, turnoutPercentage, candidates]
      properties:
        electionId:
          type: integer
        totalVotes:
          type: integer
        totalVoters:
          type: integer
        turnoutPercentage:
          type: number
        winner:
          $ref: "#/components/schemas/CandidateResult"
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/CandidateResult"

    # ─── DASHBOARD ───────────────────────────────────────────────────────────
...